move Modal component into its own folder
