+++ src/components/GameOverModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Modal from './Modal';
+import Modal from './Modal/Modal';
 import Button from './Button/Button';
 import { useResults } from '../hooks/useResults';
 import type { AnswerHistory } from '../types';
